refactor: consolidate persistent state hook
