--- src/utils/usePersistentState.ts
@@ -1,24 +1,0 @@
-import { useEffect, useState } from 'react';
-
-function usePersistentState<T>(key: string, initial: T): [T, React.Dispatch<React.SetStateAction<T>>] {
-  const [state, setState] = useState<T>(() => {
-    try {
-      const stored = localStorage.getItem(key);
-      return stored ? JSON.parse(stored) as T : initial;
-    } catch {
-      return initial;
-    }
-  });
-
-  useEffect(() => {
-    try {
-      localStorage.setItem(key, JSON.stringify(state));
-    } catch {
-      // ignore write errors
-    }
-  }, [key, state]);
-
-  return [state, setState];
-}
-
-export default usePersistentState;
